+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Removed
 ### Fixed
 
+- `EventData.Create`: restored defaulting of `EventId` to `Guid.NewGuid` broken in [#82](https://github.com/jet/FsCodec/pull/82) [#83](https://github.com/jet/FsCodec/pull/83)
+
 <a name="3.0.0-rc.6"></a>
 ## [3.0.0-rc.6] - 2022-09-02